fix issues with multiple (useless) reports generated after plugin update

+++ wwf-donations-plugin/donations/SettingsManager.php
@@ -15,17 +15,24 @@ class SettingsManager
     public const REPORT_INTERVAL_MODE_WEEKLY = 'weekly';
     public const REPORT_INTERVAL_MODE_MONTHLY = 'monthly';
     public const REPORT_INTERVAL_MODE_QUARTERLY = 'quarterly';
+    // option keys
+    const WWF_DONATIONS_REPORTING_LAST_CHECK_DATE = 'wwf_donations_reporting_last_check_date';
+    const WWF_DONATIONS_REPORTING_INTERVAL = 'wwf_donations_reporting_interval';
+    const WWF_DONATIONS_REPORTING_LIVE_DAYS_IN_PAST = 'wwf_donations_reporting_live_days_in_past';
+    const WWF_DONATIONS_REPORTING_RECIPIENT = 'wwf_donations_reporting_recipient';
+    const WWF_DONATIONS_REPORTING_LAST_GENERATION_DATE = 'wwf_donations_reporting_last_generation_date';
+    const WWF_DONATIONS_REPORTING_COUNTER = 'wwf_donations_reporting_counter';
 
     /**
      * @var array option name => default value
      */
     private static $options = [
-        'wwf_donations_reporting_interval' => self::REPORT_INTERVAL_MODE_MONTHLY,
-        'wwf_donations_reporting_live_days_in_past' => 30,
-        'wwf_donations_reporting_recipient' => 'Eshop-Spenden@wwf.de',
-        'wwf_donations_reporting_last_generation_date' => null,
-        'wwf_donations_reporting_last_check_date' => null,
-        'wwf_donations_reporting_counter' => 0,
+        self::WWF_DONATIONS_REPORTING_INTERVAL => self::REPORT_INTERVAL_MODE_MONTHLY,
+        self::WWF_DONATIONS_REPORTING_LIVE_DAYS_IN_PAST => 30,
+        self::WWF_DONATIONS_REPORTING_RECIPIENT => 'Eshop-Spenden@wwf.de',
+        self::WWF_DONATIONS_REPORTING_LAST_GENERATION_DATE => null,
+        self::WWF_DONATIONS_REPORTING_LAST_CHECK_DATE => null,
+        self::WWF_DONATIONS_REPORTING_COUNTER => 0,
     ];
 
     private static $reportingIntervalOptions = [
@@ -44,18 +51,18 @@ class SettingsManager
         }
         // ensure options are present
         foreach (self::$options as $key => $defaultValue) {
-            $optionValue = get_option($key);
-            if (!$optionValue || $optionValue !== $defaultValue) {
-                // default value
+            $optionValue = get_option($key, 'no-init');
+            if ($optionValue == 'no-init') {
+                // set default value
                 update_option($key, $defaultValue);
             }
         }
         register_setting(Plugin::$pluginSlug,
-            'wwf_donations_reporting_interval', 'esc_attr');
+            self::WWF_DONATIONS_REPORTING_INTERVAL, 'esc_attr');
         register_setting(Plugin::$pluginSlug,
-            'wwf_donations_reporting_live_days_in_past', 'esc_attr');
+            self::WWF_DONATIONS_REPORTING_LIVE_DAYS_IN_PAST, 'esc_attr');
         register_setting(Plugin::$pluginSlug,
-            'wwf_donations_reporting_recipient', 'esc_attr');
+            self::WWF_DONATIONS_REPORTING_RECIPIENT, 'esc_attr');
     }
 
     /**
@@ -63,13 +70,21 @@ class SettingsManager
      */
     public static function uninstall()
     {
+        unregister_setting(Plugin::$pluginSlug, self::WWF_DONATIONS_REPORTING_INTERVAL);
+        unregister_setting(Plugin::$pluginSlug, self::WWF_DONATIONS_REPORTING_LIVE_DAYS_IN_PAST);
+        unregister_setting(Plugin::$pluginSlug, self::WWF_DONATIONS_REPORTING_RECIPIENT);
         foreach (self::$options as $key => $defaultValue) {
             // default value = monthly
+            if (in_array($key, [
+                self::WWF_DONATIONS_REPORTING_LAST_GENERATION_DATE,
+                self::WWF_DONATIONS_REPORTING_COUNTER,
+            ])) {
+                // preserve several options, to avoid multiple reports after plugin updates
+                // should not trigger a new report: uninstall -> install -> activate -> deactivate -> activate -> repeat
+                continue;
+            }
             delete_option($key);
         }
-        unregister_setting(Plugin::$pluginSlug, 'wwf_donations_reporting_interval');
-        unregister_setting(Plugin::$pluginSlug, 'wwf_donations_reporting_live_days_in_past');
-        unregister_setting(Plugin::$pluginSlug, 'wwf_donations_reporting_recipient');
     }
 
     /**
@@ -80,8 +95,8 @@ class SettingsManager
      */
     public static function getOptionCurrentReportingInterval(): string
     {
-        $default = self::$options['wwf_donations_reporting_interval'];
-        $intervalMode = strval(get_option('wwf_donations_reporting_interval', $default));
+        $default = self::$options[self::WWF_DONATIONS_REPORTING_INTERVAL];
+        $intervalMode = strval(get_option(self::WWF_DONATIONS_REPORTING_INTERVAL, $default));
         if (!in_array($intervalMode, array_keys(self::$reportingIntervalOptions))) {
             return $default;
         }
@@ -93,8 +108,8 @@ class SettingsManager
      */
     public static function getOptionLiveReportDaysInPast(): int
     {
-        return intval(get_option('wwf_donations_reporting_live_days_in_past',
-            self::$options['wwf_donations_reporting_live_days_in_past']));
+        return intval(get_option(self::WWF_DONATIONS_REPORTING_LIVE_DAYS_IN_PAST,
+            self::$options[self::WWF_DONATIONS_REPORTING_LIVE_DAYS_IN_PAST]));
     }
 
     /**
@@ -102,8 +117,8 @@ class SettingsManager
      */
     public static function getOptionReportRecipientMail(): string
     {
-        return strval(get_option('wwf_donations_reporting_recipient',
-            self::$options['wwf_donations_reporting_recipient']));
+        return strval(get_option(self::WWF_DONATIONS_REPORTING_RECIPIENT,
+            self::$options[self::WWF_DONATIONS_REPORTING_RECIPIENT]));
     }
 
     /**
@@ -111,8 +126,8 @@ class SettingsManager
      */
     public static function getOptionReportLastGenerationDate(): ?\DateTime
     {
-        $storedValue = strval(get_option('wwf_donations_reporting_last_generation_date',
-            self::$options['wwf_donations_reporting_last_generation_date']));
+        $storedValue = strval(get_option(self::WWF_DONATIONS_REPORTING_LAST_GENERATION_DATE,
+            self::$options[self::WWF_DONATIONS_REPORTING_LAST_GENERATION_DATE]));
         if (!$storedValue) {
             return null;
         }
@@ -136,13 +151,13 @@ class SettingsManager
                 $dateTime = strtotime('now');
             }
         }
-        update_option('wwf_donations_reporting_last_generation_date', $dateTime->format('c'));
+        update_option(self::WWF_DONATIONS_REPORTING_LAST_GENERATION_DATE, $dateTime->format('c'));
     }
 
     public static function getOptionReportLastCheck(): ?\DateTime
     {
-        $storedValue = strval(get_option('wwf_donations_reporting_last_check_date',
-            self::$options['wwf_donations_reporting_last_check_date']));
+        $storedValue = strval(get_option(self::WWF_DONATIONS_REPORTING_LAST_CHECK_DATE,
+            self::$options[self::WWF_DONATIONS_REPORTING_LAST_CHECK_DATE]));
         if (!$storedValue) {
             return null;
         }
@@ -159,7 +174,7 @@ class SettingsManager
      */
     public static function setOptionReportLastCheck(): void
     {
-        update_option('wwf_donations_reporting_last_check_date', date('c'));
+        update_option(self::WWF_DONATIONS_REPORTING_LAST_CHECK_DATE, date('c'));
     }
 
     /**
@@ -169,8 +184,8 @@ class SettingsManager
      */
     public static function getOptionReportCounterIncremented(): int
     {
-        $incrementedCounter = intval(get_option('wwf_donations_reporting_counter', 0)) + 1;
-        update_option('wwf_donations_reporting_counter', $incrementedCounter);
+        $incrementedCounter = intval(get_option(self::WWF_DONATIONS_REPORTING_COUNTER, 0)) + 1;
+        update_option(self::WWF_DONATIONS_REPORTING_COUNTER, $incrementedCounter);
         return $incrementedCounter;
     }
